+++ lua/snippets/luasnip/init.lua
@@ -1,3 +1,4 @@
 require('snippets.luasnip.go')
 require('snippets.luasnip.lua_snippets')
+require('snippets.luasnip.markdown')
 require('luasnip.loaders.from_vscode').lazy_load()