+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 > Features
 - **luasnip:** added virtual text on choice nodes
 - **mappings:** moved q to <leader>q
+- **snippet:** added table generating snippets markdown lua
 - **snippet:** added dynamic req snippet
 - **snippet:** added prequire snippet lua
 - **snippet:** more snippet for go